requirements.txt
fastapi~=0.111.0
uvicorn~=0.30.1
jose~=1.0.0
python-multipart~=0.0.9
bcrypt~=4.0.0
passlib~=1.7.4
pydantic~=2.7.1
PyJWT~=2.8.0
passlib~=1.7.4
requests~=2.32.2
sendgrid~=6.11.0
beanie~=1.26.0
pymongo~=4.7.3
motor~=3.4.0